bullet position fix

+++ Bullet.cpp
@@ -2,9 +2,9 @@
 
 Bullet::Bullet() : position(0, 0) {}
 
-Bullet::Bullet(int x_, int y_, sf::Texture &texture_) : position(x_, y_), sprite(texture_) { sprite.setPosition((float)x_, (float)y_); }
+Bullet::Bullet(float x_, float y_, sf::Texture &texture_) : position(x_, y_), sprite(texture_) { sprite.setPosition((float)x_, (float)y_); }
 
-Point<int> Bullet::getPosition() const { return position; }
+Point<float> Bullet::getPosition() const { return position; }
 
 sf::Sprite Bullet::getSprite() const { return sprite; }
 
@@ -14,6 +14,6 @@ std::ostream &operator<<(std::ostream &os, const Bullet &bullet) {
 }
 
 void Bullet::move(int direction) {
-    position.setY(position.getY() + direction * 2);
+    position.setY(position.getY() + (float)direction * 2.0f);
     sprite.setPosition((float)position.getX(), (float)position.getY());
 }
